Fix: Hook into `emit` instead of `afterEmit`
Newer versions of webpack will not allow access to the sources of an asset after it has been emitted (to save memory while building).

This commit describes the new logic, which can be switched on with the `futureEmitAssets` option:
https://github.com/webpack/webpack/commit/aaf85dbd1c221eb1788dd46894d24ebd9c62aa62

The new webpack logic breaks the plugin which assumes the source is still accessible at the `afterEmit` stage, resulting in an exception while uploading (`”Content and Map of this Source is no longer available (only size() is supported)”`).

As a fix, this uses the `emit` instead of the `afterEmit` hook.